[Lab/LDAP] Fix user create instructions, add notes
The new version of the 389ds cockpit extension does not use the correct
classes to be able to set a user's `cn` and `sn` when using the
"create a new user" option. Fix this by using the custom entry
option instead.

Also improve the readability by adding a note for the SSH SOCKS proxy
and changing the format of the SSSD config file to highlight that the
directory manager's password must be used when authenticating as the
directory manager.